chore: swap flaky current chats e2e tests for unit tests
